Sanketika-Obsrv/issue-tracker#154: switch for generating different ingestion template
